Use busy-wait I/O (no interrupts) in "Test 2"
Interrupts were an inadvertent test actor in Test 2.  But Test 2 is
supposed to merely monitor timing, with no early wake-ups from tickless
idle.  Note that we still use interrupts in Test 3, which is supposed
to stress the tickless logic via early wake-ups.

After this commit, the default tickless idle really does have trivial
drift in Test 2, as already noted in the test results in ReadMe.md.